# Refactor Batch Processor benchmark to really test OnEnd