@pellared 2024-07-15
@open-telemetry/go-approvers, this needs a non-Splunk employee approval.